% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{SF6D_distributions}
\alias{SF6D_distributions}
\alias{SF12_to_PF}
\alias{SF12_to_RL}
\alias{SF12_to_SF}
\alias{SF12_to_PAIN}
\alias{SF12_to_MH}
\alias{SF12_to_VIT}
\alias{SF12_to_MOST}
\title{Predict sample distribution of SF-6D dimensions}
\usage{
SF12_to_PF(means, SDs, corr = NULL, version = 2)

SF12_to_RL(means, SDs, corr = NULL, version = 2)

SF12_to_SF(means, SDs, corr = NULL, version = 2)

SF12_to_PAIN(means, SDs, corr = NULL, version = 2)

SF12_to_MH(means, SDs, corr = NULL, version = 2)

SF12_to_VIT(means, SDs, corr = NULL, version = 2)

SF12_to_MOST(means, SDs, corr = NULL, version = 2)
}
\arguments{
\item{means}{A vector: \code{(mean(PCS), mean(MCS))}}

\item{SDs}{A vector: \code{(sd(PCS), sd(MCS))}}

\item{corr}{\strong{(optional)} A matrix: \code{(Cor(PCS, MCS))}. Default is to use the correlation matrix obtained
from the estimation sample.}

\item{version}{\strong{(optional)} An integer: 1 = PCS/MCS scores are from SF-12-v1; 2 (default) = PCS/MCS scores are
from SF-12-v2. Other values give an error.}
}
\value{
\code{SF12_to_PF}: A vector of length three: \code{(Pr(PF==1), Pr(PF==2), Pr(PF==3))}

\code{SF12_to_PF}: A vector of length two: \code{(Pr(RL==1), Pr(PF>1))}

\code{SF12_to_PF}: A vector of length five: \code{(Pr(SF==1), ..., Pr(SF==5))}

\code{SF12_to_PF}: A vector of length five: \code{(Pr(PAIN==1), ..., Pr(PAIN==5))}

\code{SF12_to_PF}: A vector of length five: \code{(Pr(MH==1), ..., Pr(MH==5))}

\code{SF12_to_PF}: A vector of length five: \code{(Pr(VIT==1), ..., Pr(VIT==5))}

\code{SF12_to_PF}: A vector of length two: \code{(Pr(MOST==0), Pr(MOST==1))}
}
\description{
These functions provide estimates of the predicted sample distribution of each SF-6D dimension from the sample means
  and standard deviations of SF-12 PCS and MCS scores (and optionally their correlation).
}
\seealso{
\code{\link{SF12_to_SF6D}} for calculating the resulting utility values.
}
