% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{SF6D_uncertainty}
\alias{SF6D_uncertainty}
\title{Estimate prediction uncertainty in mapped mean SF-6D utility values}
\usage{
SF6D_uncertainty(N, dimensions = NULL, level = 0.95, version = 2)
}
\arguments{
\item{N}{A positive integer: sample size.}

\item{dimensions}{\strong{(optional)} A character vector: one or more of "utility", "PF", "RL", "SF", "PAIN", "MH",
"VIT", "MOST", indicating which dimensions should be predicted ("utility" (the default) gives the overall SF-6D
utility score); or "all", indicating all of the above.}

\item{level}{\strong{(optional)} A numeric scalar in (0, 1): significance level for uncertainty intervals. Default is
0.95, for 95 percent uncertainty intervals. Can also be specified as \code{"se"} to give standard errors
(equivalent to \code{level=1-2*(1-pnorm(1))}).}

\item{version}{\strong{(optional)} An integer: 1 = PCS/MCS scores are from SF-12-v1; 2 (default) = PCS/MCS scores are
from SF-12-v2. Other values give an error.}
}
\value{
A numeric vector of length \code{length(dimensions)} (or length 8 for \code{dimensions == "all"}):
  (half-)width of the estimated uncertainty intervals for sample mean utility and/or contributions to mean utility of
  each dimension (i.e. (level * 100)\% uncertainty interval = point estimate ± return value)
}
\description{
\code{SF6D_uncertainty} provides an estimate of the prediction uncertainty associated with the mean utility value
  predicted by the mapping algorithm.
}
\seealso{
\code{\link{SF6D_distributions}} for the dimension-specific functions, and \code{\link{SF12_to_SF6D}} for
the estimated sample mean utility values.
}
